Fix #1968
